rename logger.rb to logging.rb
Having `require 'logger'` inside logger.rb became confusing when minitest changed load order.
