+++ api/client.rb
@@ -1,6 +1,6 @@
 # frozen_string_literal: true
 
-require_relative '../logger'
+require_relative '../logging'
 
 class APIClient
   include Loggable
