+++ logging.rb
